+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "Background Phi Colors" extension will be documented 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 0.?.? - 2019-0?-??
+
+### Changed
+
+- Renewal of extension icon.
+
 ## 0.4.0 - 2019-03-26
 
 ### Added