+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "Background Phi Colors" extension will be documented 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 0.6.0 - 2019-04-??
+
+### Added
+
+- Active Scope
+
 ## 0.5.0 - 2019-04-02
 
 ### Added